refactor(cli): complete orchestration extraction; close backlog item
- Move ExecutionMode + impls to exec_profile.rs (removes circular super:: import)
- Move SolverMode, PulseRhsMode, BenchRecord, and solver constants (C0,
  CONTINUITY_REL_RESIDUAL_MAX, SINUSOIDAL_REL_RESIDUAL_MAX) to solve_session.rs
- Extract BenchFormat + all bench-emit helpers into new bench.rs module
- Extract hybrid-parallel sweep dispatch into execute_frequency_sweep() in
  solve_session.rs; main() delegates with a single call and handles warnings
- Update cli_args.rs and warnings.rs imports to reference the new home modules
- main.rs reduced from 628 → 393 lines; fn main() is now frontend wiring only

Resolution: backlog item 'CLI orchestration extraction' closed. All six
extracted modules are independently importable and testable without spawning
the CLI binary. All tests pass.
